refactor the project structure
